requirements.txt
# Streamlit for the UI
streamlit

# Search Engine dependencies
pandas
numpy
sentence-transformers
torch
faiss-cpu